Information criterion computer

Computes information criterion based on Renyi entropy for determining the size of the logos in the logo plot. Similar but allows for more general entropy criteria to the pwm2ic() function in seqLogo package. By tuning the alpha parameter input, one can vary the entropy from Shannon (alpha=1) to collision (alpha=2) to min-entropy (alpha=infty)

ic_computer(mat, alpha)

Arguments

mat

A matrix with symbols or logo names along the rows and the sites/positions/groups along the columns.

alpha

The tuning parameter of the Renyi entropy used in computing the information criterion. Default is alpha=1, for which it uses Shannon entropy (in the limit).

Value

A vector of same length as the number of columns in the data, with each entry representing information contained in that column, which determines the height of the bar for the logo plot for that particular column (site/position/block).

Examples

counts_mat <- rbind(c(0, 10, 100, 60, 20),
                   c(40, 30, 30, 35, 20),
                   c(100, 0, 15, 25, 75),
                   c(10, 30, 20, 50, 70)
              )
colnames(counts_mat) <- c("2012", "2013", "2014", "2015", "2016")
rownames(counts_mat) <- c("P1", "P2", "P3", "P4")
ic_computer(counts_mat, alpha=2)

#> [1] 1.3460735 1.0526187 1.1403829 0.7093303 0.8940563